:recycle: refactor: Refactor get all product
