+++ src/main/kotlin/com/highv/ecommerce/domain/product/controller/ProductController.kt
@@ -1,7 +1,5 @@
 package com.highv.ecommerce.domain.product.controller
 
-import com.highv.ecommerce.domain.backoffice.dto.productbackoffice.ProductBackOfficeRequest
-import com.highv.ecommerce.domain.product.dto.CreateProductRequest
 import com.highv.ecommerce.domain.product.dto.CreateRequest
 import com.highv.ecommerce.domain.product.dto.ProductResponse
 import com.highv.ecommerce.domain.product.dto.UpdateProductRequest
@@ -14,8 +12,14 @@ import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.security.core.annotation.AuthenticationPrincipal
-import org.springframework.web.bind.annotation.*
-import org.springframework.web.multipart.MultipartFile
+import org.springframework.web.bind.annotation.DeleteMapping
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PatchMapping
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
 
 @RestController
 @RequestMapping("/api/v1/products")
@@ -64,13 +68,6 @@ class ProductController(
     ): ResponseEntity<ProductResponse> = ResponseEntity
         .status(HttpStatus.OK)
         .body(productService.getProductById(productId))
-
-    @GetMapping("/all")
-    fun getAllProducts(
-        @PageableDefault(size = 10, page = 0) pageable: Pageable
-    ): ResponseEntity<Page<ProductResponse>> = ResponseEntity
-        .status(HttpStatus.OK)
-        .body(productService.getAllProducts(pageable))
 
     @GetMapping("/category")
     fun getProductsByCategory(
