+++ src/main/kotlin/com/highv/ecommerce/domain/product/controller/ProductSearchController.kt
@@ -5,7 +5,9 @@ import com.highv.ecommerce.domain.product.dto.TopSearchKeyword
 import com.highv.ecommerce.domain.product.service.ProductSearchService
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.PageRequest
+import org.springframework.data.domain.Pageable
 import org.springframework.data.domain.Sort
+import org.springframework.data.web.PageableDefault
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
@@ -38,4 +40,11 @@ class ProductSearchController(
             .status(HttpStatus.OK)
             .body(productSearchService.searchByRedis(keyword, pageRequest))
     }
+
+    @GetMapping("/all")
+    fun getAllProducts(
+        @PageableDefault(size = 10, page = 0) pageable: Pageable
+    ): ResponseEntity<Page<ProductResponse>> = ResponseEntity
+        .status(HttpStatus.OK)
+        .body(productSearchService.getAllProducts(pageable))
 }
